refactored a little more to keep code clean
