+++ lib/deleteHandler.js
@@ -11,7 +11,7 @@ function deleteHandler(parsedPathObject, deleteResponse) {
                 if (err) deleteResponse(err);
                 else deleteResponse(null,'File was deleted');
             });
-        }
+        };
     });
 };
