+++ lib/getHandler.js
@@ -10,14 +10,11 @@ function getHandler(parsedPathObject, getResponse) {
         sander
             .readdir( resourceDir )
             .then( data => {
-                if (parsedName === 'cats') {
-                    getResponse(null, data.toString(), 'text/plain');
-                } else {
-                    return sander
-                            .readFile(`${resourceDir}${parsedName}.json`, {'encoding': 'utf-8'})
-                            .then(data => { getResponse(null, data, 'application/json'); })
-                            .catch(err => { getResponse(err); });
-                };
+                if (parsedName === 'cats') getResponse(null, data.toString(), 'text/plain');
+                else sander
+                    .readFile(`${resourceDir}${parsedName}.json`, {'encoding': 'utf-8'})
+                    .then(data => { getResponse(null, data, 'application/json'); })
+                    .catch(err => { getResponse(err); });
             })
             .catch(err => { getResponse(err); });
     } else {
