+++ lib/httpServer.js
@@ -20,7 +20,7 @@ const server = http.createServer((req, res) => {
 
     if (req.method === 'GET') {
         getHandler(parsedPath, (err, data, encoding) => {
-            if (err) resHandler(err, 400);
+            if (err) resHandler(err.code, 400);
             else resHandler(data, null, encoding);                
         });
     } else if (req.method === 'PUT') {
