+++ lib/postHandler.js
@@ -20,7 +20,7 @@ function postHandler(parsedPathObject, req, postResponse) {
                 });
                 req.on('end', () => {
                     dataId = array[(array.length - 1)] + 1;
-                    return sander
+                    sander
                         .writeFile(`${resourceDir}${dataId}.json`, body)
                         .then(data => { postResponse(null, 'post good, your resource id is ' + dataId); })
                         .catch(err => { postResponse(err); });
